api/pyproject.toml
[project]
name = "physical-ai-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agents>=1.4.0",
    "bcrypt==4.0.1",
    "cryptography>=46.0.3",
    "fastapi[standard]>=0.122.0",
    "fastembed>=0.4.2",
    "httpx>=0.28.1",
    "openai>=1.0.0",
    "openai-agents>=0.6.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "qdrant-client>=1.12.1",
    "sse-starlette>=2.1.3",
    "tiktoken>=0.7.0",
    "uvicorn>=0.33.0",
]
